This is my initial draft of the aMSN2 project. It has so far a good MSNP protocol API with a very good (I like to hope) design which will make implementing new protocols VERY easy in the future... please all of you.. read this code, start with protocols/msnp/msnp.tcl and read through the rest, see how the delegation and heritage (through the $protocol variable call) works (msnp15.tcl -> msnp14.tcl -> msnp13.tcl -> msnp12.tcl -> msnp11.tcl -> msnp.tcl -> protocolapi.tcl) and how easy it is to implement new protocol versions...
There is only ONE account manager created by default, and stored in $::accountManager 
To test this, launch amsn and type :
  reload_amsn2
  $::accountManager Connect
then follow it up in the status log (not protocol log yet).
to disconnect...
  $::accountManager DisConnect
no other API is available 'yet'.
The contact list should load and you should get your users and groups and the nickname/psm/status of each user and it should work for a little bit without any problems... 
It will use your config and will connect using your currently selected profile (you can't test on default profile I guess...)

remaining in MSNP11 : add/remove/rename of groups and contacts + handling of the SB

The whole protocol is totally separated from the rest, we have 4 main parts :
the protocol
the core -> accountmanager
the core -> datamanger
the gui -> using amsn1 for now...

the datamanager is used to handle the data associated with an account (would replace abook) and is completely separated from all the other parts, it knows nothing about the other layers.
the accountmanager is a singleton (only one instance) and should allow us to connect to each account, it also serves as a layer in between the protocol, the data and the gui, it catches protocol events, updates the associated data, then sends (or should send) a new event to the gui. protocol and gui never accesses the accountmanager directly, they always have to send an event to communicate with the accoutmanager
the protocol and the gui are completely separate, they both have access to the datamanager but ONLY in readonly mode.
For example, if the protocol for example ever wants to change something in the data (a user changed his status) it won't change it from the datamanager, it would instead send an event notifying the core of the new status, the accountmanager would catch the event, and modify the datamanager, then resend the event to the gui layer (as a DataChangedNotification), the gui will then read the new data from the datamanager and update itself. 
in short :
datamanager : knows noone
accountmanager : has WRITE access to the datamanager, catches and sends events to/from the protocol and gui layers. Can access the protocol/gui layers, but only as the 'caller object' from an event. It will not store the objects as a snit variable.
protocol : has READ access to the datamanager and sends events to the accountmanager
gui : has READ access to the datamanager and sends events to the accountmanager


many things are still TODO... don't touch anything without asking first :@ I don't want the whole design to crumble... :@